# Improve performance of FASTA random reading