--- a/.travis.yml
+++ b/.travis.yml
@@ -1,5 +1,5 @@
 language: clojure
-lein: 2.8.0
+lein: 2.8.1
 cache:
   directories:
     - $HOME/.m2