#### Summary
Improve performance of FASTA random reading by **10x** using MappedByteBuffer and CharBuffer.

#### Changes
Current implementation of `read-sequence` for FASTA is slow because of multiple String generation.
byte array => String => String => String

This PR uses mmap and CharBuffer to generate String once.
MappedByteBuffer => CharBuffer => String

#### Benchmark
##### benchmarking code
```clojure
(are [reg opts] (c/quick-bench (with-open [rdr (cseq/reader "/path/to/hg38.fa")]
                                 (dorun (cseq/read-sequence rdr reg opts))))
  {:chr "chr1" :start 1 :end 248956422} {:mask? true}
  {:chr "chr1" :start 1 :end 248956422} {})
```

##### master
```
[{:chr "chr1", :start 1, :end 248956422} {:mask? true}]
time: 20.807584 sec, sd: 23.176079 ms
[{:chr "chr1", :start 1, :end 248956422} {}]
time: 22.297717 sec, sd: 30.394828 ms
```

##### this branch
```
[{:chr "chr1", :start 1, :end 248956422} {:mask? true}]
time: 1.994279 sec, sd: 9.574159 µs
[{:chr "chr1", :start 1, :end 248956422} {}]
time: 2.007051 sec, sd: 312.275365 µs
```

#### Tests

- `lein test :all` 🆗