@totakke 2017-11-10
Although not related to this change, I felt the way of handling external region is somewhat strange.

When the supplied region is partially over the sequence, missing bases are filled with `N`.

```clojure
(with-open [rdr (reader "test-resources/fasta/test.fa")]
  (read-sequence rdr {:chr "ref" :start 30 :end 60}))
;;=> AGGCAGTCAGCGCCATNNNNNNNNNNNNNNN
```

But when the region is not included in sequence, `read-sequence` returns `nil`.

```clojure
(with-open [rdr (reader "test-resources/fasta/test.fa")]
   (read-sequence rdr {:chr "ref" :start 50 :end 60}))
;;=> nil
```

This behavior seems inconsistent. What do you think about this?