@alumi 2017-11-10
You're absolutely right!
I think we don't need padding because no other readers do pad anything.
I'll create a PR for this so please check it later.

Thank you.